Use a more sensible type for DBAPI connection reuse counts.  JIRA: CXX-4373.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@66108 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/dbapi/driver/impl/dbapi_impl_connection.hpp
@@ -265,7 +265,7 @@ protected:
     void CheckCanOpen(void);
     void MarkClosed(void);
 
-    size_t GetReuseCount(void) const
+    unsigned int GetReuseCount(void) const
     {
         return m_ReuseCount;
     }
@@ -314,7 +314,7 @@ private:
     const Uint2    m_Port;
     const string   m_Passwd;
     const string   m_Pool;
-    size_t         m_ReuseCount;
+    unsigned int   m_ReuseCount;
     const bool     m_Reusable;
     bool           m_OpenFinished;
     bool           m_Valid;
